refactor: move non-route export: partitionRanges
