+++ bad-exports.txt
@@ -1,0 +1,13 @@
+[error:bad-export] app/routes/decks/$id/index.tsx:67 QueueStatisticsComponent
+[error:bad-export] app/routes/decks/$id/index.tsx:127 PRACTICE_ACTION_TYPE_TO_COLOR
+[error:bad-export] app/routes/decks/$id/index.tsx:217 QueueTypeIcon
+[error:bad-export] app/routes/decks/$id/index.tsx:243 DeckNavBarMenuComponent
+[error:bad-export] app/routes/decks/$id/index.tsx:248 DeckMenuComponent
+[error:bad-export] app/routes/decks/$id/history-graph.tsx:145 DeckHistoryNavBarMenuComponent
+[error:bad-export] app/routes/videos/index.tsx:38 VideoListComponent
+[error:bad-export] app/routes/videos/$id.utils.tsx:3 partitionRanges
+[error:bad-export] app/routes/bookmarks/index.tsx:113 BookmarkEntryComponent
+[error:bad-export] app/routes/bookmarks/index.tsx:185 MiniPlayer
+[error:bad-export] app/routes/bookmarks/index.tsx:430 BookmarksMenuItems
+[error:bad-export] app/routes/videos/$id.tsx:60 findCurrentEntry
+[error:bad-export] app/routes/videos/$id.tsx:510 CaptionEntryComponent
